+++ src/model/table/tableNode.ts
@@ -63,7 +63,7 @@ export class TableNode implements INode, IConnection {
 
         vscode.window.showInputBox({ value: this.table, placeHolder: 'newTableName', prompt: `You will changed ${this.database}.${this.table} to new table name!` }).then(async newTableName => {
             if (!newTableName) return
-            const sql = `alter table ${this.database}.${this.table} rename ${newTableName}`
+            const sql = `RENAME TABLE \`${this.database}\`.\`${this.table}\` to \`${this.database}\`.\`${newTableName}\``
             QueryUnit.queryPromise(await ConnectionManager.getConnection(this), sql).then((rows) => {
                 DatabaseCache.clearTableCache(`${this.host}_${this.port}_${this.user}_${this.database}`)
                 MySQLTreeDataProvider.refresh()
@@ -76,8 +76,9 @@ export class TableNode implements INode, IConnection {
     public dropTable() {
 
         vscode.window.showInputBox({ prompt: `Are you want to drop table ${this.table} ?     `, placeHolder: 'Input y to confirm.' }).then(async inputContent => {
+            if(!inputContent)return;
             if (inputContent.toLocaleLowerCase() == 'y') {
-                QueryUnit.queryPromise(await ConnectionManager.getConnection(this), `DROP TABLE ${this.database}.${this.table}`).then(() => {
+                QueryUnit.queryPromise(await ConnectionManager.getConnection(this), `DROP TABLE \`${this.database}\`.\`${this.table}\``).then(() => {
                     DatabaseCache.clearTableCache(`${this.host}_${this.port}_${this.user}_${this.database}`)
                     MySQLTreeDataProvider.refresh()
                     vscode.window.showInformationMessage(`Drop table ${this.table} success!`)
@@ -93,8 +94,9 @@ export class TableNode implements INode, IConnection {
     truncateTable() {
 
         vscode.window.showInputBox({ prompt: `Are you want to clear table ${this.table} all data ?          `, placeHolder: 'Input y to confirm.' }).then(async inputContent => {
+            if(!inputContent)return;
             if (inputContent.toLocaleLowerCase() == 'y') {
-                QueryUnit.queryPromise(await ConnectionManager.getConnection(this), `truncate table ${this.database}.${this.table}`).then(() => {
+                QueryUnit.queryPromise(await ConnectionManager.getConnection(this), `truncate table \`${this.database}\`.\`${this.table}\``).then(() => {
                     vscode.window.showInformationMessage(`Clear table ${this.table} all data success!`)
                 })
             }
@@ -105,8 +107,8 @@ export class TableNode implements INode, IConnection {
 
     indexTemplate() {
         ConnectionManager.getConnection(this, true)
-        QueryUnit.showSQLTextDocument(`-- ALTER TABLE ${this.database}.${this.table} DROP INDEX [indexName];
--- ALTER TABLE ${this.database}.${this.table} ADD [UNIQUE|KEY|PRIMARY KEY] INDEX ([column]);`)
+        QueryUnit.showSQLTextDocument(`-- ALTER TABLE \`${this.database}\`.\`${this.table}\` DROP INDEX [indexName];
+-- ALTER TABLE \`${this.database}\`.\`${this.table}\` ADD [UNIQUE|KEY|PRIMARY KEY] INDEX ([column]);`)
         setTimeout(() => {
             QueryUnit.runQuery(`SELECT column_name,table_schema,index_name,non_unique FROM INFORMATION_SCHEMA.STATISTICS WHERE table_schema='${this.database}' and table_name='${this.table}';`, this)
         }, 10);
@@ -115,7 +117,7 @@ export class TableNode implements INode, IConnection {
 
 
     public async selectSqlTemplate(run: Boolean) {
-        const sql = `SELECT * FROM ${this.database}.${this.table} LIMIT ${Constants.DEFAULT_SIZE};`;
+        const sql = `SELECT * FROM \`${this.database}\`.\`${this.table}\` LIMIT ${Constants.DEFAULT_SIZE};`;
 
         if (run) {
             ConnectionManager.getConnection(this, true)
@@ -131,10 +133,10 @@ export class TableNode implements INode, IConnection {
             .getChildren()
             .then((children: INode[]) => {
                 const childrenNames = children.map((child: any) => child.column.COLUMN_NAME);
-                let sql = `insert into ${this.database}.${this.table}\n`
+                let sql = `insert into ${this.database}.{this.table}\n`
                 sql += `(${childrenNames.toString().replace(/,/g, ", ")})\n`
                 sql += "values\n"
-                sql += `(${childrenNames.toString().replace(/,/g, ", ")});`
+                sql += `('${childrenNames.toString().replace(/,/g, ", ")}');`
                 QueryUnit.createSQLTextDocument(sql);
             });
     }
@@ -161,7 +163,7 @@ export class TableNode implements INode, IConnection {
                 const childrenNames = children.filter((child: any) => !child.column.COLUMN_KEY).map((child: any) => child.column.COLUMN_NAME);
 
                 const sets = childrenNames.map((name: string) => `${name} = ${name}`);
-                const where = keysNames.map((name: string) => `${name} = ${name}`);
+                const where = keysNames.map((name: string) => `${name} = '${name}'`);
 
                 let sql = `update ${this.database}.${this.table} \nset ${sets.toString().replace(/,/g, "\n  , ")}\n`;
                 sql += `where ${where.toString().replace(/,/g, "\n   and ")}`
